+++ src/Symfony/Bundle/SecurityBundle/CHANGELOG.md
@@ -21,6 +21,7 @@ CHANGELOG
  * Add `prompt`, `max_age` and `authorization_params` options to the `oidc_login` authenticator
  * Add `user_data_source` and `user_identifier_claim` options to the `oidc_login` authenticator
  * Add `token_endpoint_auth_method` option (`client_secret_post`, `client_secret_basic`) to the `oidc_login` authenticator
+ * Add `enable_end_session` and `post_logout_redirect_path` options to the `oidc_login` authenticator for RP-Initiated Logout
 
 8.1
 ---